Backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
langgraph>=0.0.40
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-community>=0.0.10
pydantic>=2.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
typing-extensions>=4.8.0
aiohttp>=3.9.1
google-generativeai>=0.3.2
chromadb>=0.4.22
sentence-transformers>=2.2.2
langchain_chroma>=0.2.5
langchain-huggingface>=0.3.1
Pillow>=10.0.0
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
ftfy>=6.1.1
regex>=2023.8.8 